Paid Memberships Pro - Subscription Check Add On
This modification should make the plugin work with the new version.

This plugin has not been updated since 2 years ago and when we click Start it doesn't find anything because new gateways have been added and their names are different now. I presume you are using PayPal Express as your gateway like me. So I edited the code and now it is working. Please check what I changed in the plugin